TeachPendant/mcp_servers/fanuc_knowledge/pyproject.toml
[project]
name = "fanuc_knowledge"
version = "0.1.0"
description = "MCP server for semantic + keyword search over the TeachPendant FANUC normalized dataset"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mcp>=0.9.0",
  "pyyaml>=6.0.1",
  "httpx>=0.27.0"
]

[project.optional-dependencies]
embeddings = [
  "numpy>=1.26",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fanuc_knowledge"]
